make it possible to build for windows
This commit introduces a script and instructions for building Tella
Desktop for Windows. This commit also changes the structure of the
database migrations to circumvent the golang `embed` language directive,
which seemed to not work in the otherwise functional Windows build.

Instead of embedding a plaintext file for migrations inside the binary,
we put its plaintext contents inside an array that is returned by a util
function as part of the database package.